# Improve performance of analyzer driver in presence of symbol start an…