@AlekseyTs 2020-03-23
@dotnet/roslyn-compiler For the second review.